feat: add native audio foundation
Extract system input and Android decode helpers into native audio modules.

Add fallback command shell and include Tauri shell tests in local and CI flows.
