+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -28,7 +28,7 @@
 - [ ] PR title follows [Conventional Commits](https://www.conventionalcommits.org/) (e.g. `feat(backend): add chord smoothing window`) — it becomes the squash commit message on `main`
 - [ ] `pnpm lint` passes
 - [ ] `pnpm typecheck` passes
-- [ ] `pnpm test` passes (desktop + backend)
+- [ ] `pnpm test` passes (desktop + Tauri shell + backend)
 - [ ] If backend routes or schemas changed, ran `pnpm contracts:generate` and committed the result
 - [ ] Updated relevant docs (README, backend README, THIRD_PARTY_NOTICES) if behavior changed
 - [ ] No secrets, credentials, or copyrighted audio added to the repository
