+++ .github/workflows/ci.yml
@@ -115,6 +115,22 @@ jobs:
         with:
           python-version: "3.11"
 
+      - name: Set up Rust
+        run: |
+          rustup toolchain install stable --profile minimal
+          rustup default stable
+
+      - name: Install Tauri system dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y \
+            libwebkit2gtk-4.1-dev \
+            libgtk-3-dev \
+            libayatana-appindicator3-dev \
+            librsvg2-dev \
+            libxdo-dev \
+            libssl-dev
+
       - name: Install dependencies
         run: pnpm install --frozen-lockfile
 
@@ -136,3 +152,11 @@ jobs:
 
       - name: Test desktop
         run: pnpm --filter @tuneforge/desktop test --run
+
+      - name: Check Tauri shell
+        working-directory: apps/desktop/src-tauri
+        run: cargo check
+
+      - name: Test Tauri shell
+        working-directory: apps/desktop/src-tauri
+        run: cargo test
