+++ apps/desktop/src-tauri/src/mobile_backend.rs
@@ -249,16 +249,16 @@ mobile_stub!(mobile_cancel_job, JobResponse, app: AppHandle, job_id: String);
 #[cfg(target_os = "android")]
 mod android {
     use super::*;
+    use crate::native_audio::decode::{
+        read_mobile_audio, read_resampled_mono_audio, write_mono_pcm_wav,
+    };
     use android_system_properties::AndroidSystemProperties;
     use chrono::{SecondsFormat, Utc};
     use rusqlite::{params, Connection, OptionalExtension, Row};
     use serde_json::json;
     use std::{
-        ffi::{CStr, CString},
-        fs, io, mem,
-        os::{fd::AsRawFd, raw::c_char},
+        fs, io,
         path::{Path, PathBuf},
-        ptr, slice,
         str::FromStr,
         thread,
         time::Instant,
@@ -280,12 +280,6 @@ mod android {
         path: PathBuf,
         name: &'static str,
         max_recommended_model: &'static str,
-    }
-
-    struct DecodedAudio {
-        samples: Vec<f32>,
-        sample_rate: u32,
-        channels: u32,
     }
 
     struct MobileLyricsTranscription {
@@ -768,17 +762,6 @@ mod android {
         pitch_classes: [f64; 12],
     }
 
-    fn read_mobile_audio(path: &Path) -> Result<DecodedAudio, String> {
-        match read_wav_audio(path) {
-            Ok(audio) => Ok(audio),
-            Err(wav_error) => read_android_media_audio(path).map_err(|media_error| {
-                format!(
-                    "Mobile audio decode supports PCM WAV files and formats Android MediaCodec can decode. WAV decode failed: {wav_error}. Android MediaCodec decode failed: {media_error}"
-                )
-            }),
-        }
-    }
-
     fn read_audio_features(path: &Path) -> Result<MobileAudioFeatures, String> {
         const MAX_ANALYSIS_SECONDS: usize = 30;
         let audio = read_mobile_audio(path)?;
@@ -901,700 +884,6 @@ mod android {
             attach_playback_proxy_metadata(connection, &source_artifact.id, &playback_path)?;
         }
         Ok(())
-    }
-
-    fn write_mono_pcm_wav(path: &Path, audio: &DecodedAudio) -> Result<(), String> {
-        if audio.sample_rate == 0 {
-            return Err("Decoded audio contained invalid stream metadata.".to_string());
-        }
-        let data_bytes = audio
-            .samples
-            .len()
-            .checked_mul(2)
-            .ok_or_else(|| "Decoded audio is too large for a playback proxy.".to_string())?;
-        let riff_size = 36usize
-            .checked_add(data_bytes)
-            .ok_or_else(|| "Decoded audio is too large for a playback proxy.".to_string())?;
-        let data_bytes = u32::try_from(data_bytes)
-            .map_err(|_| "Decoded audio is too large for a playback proxy.".to_string())?;
-        let riff_size = u32::try_from(riff_size)
-            .map_err(|_| "Decoded audio is too large for a playback proxy.".to_string())?;
-        let byte_rate = audio
-            .sample_rate
-            .checked_mul(2)
-            .ok_or_else(|| "Decoded audio sample rate is invalid.".to_string())?;
-
-        let mut bytes = Vec::with_capacity(44 + data_bytes as usize);
-        bytes.extend_from_slice(b"RIFF");
-        bytes.extend_from_slice(&riff_size.to_le_bytes());
-        bytes.extend_from_slice(b"WAVE");
-        bytes.extend_from_slice(b"fmt ");
-        bytes.extend_from_slice(&16u32.to_le_bytes());
-        bytes.extend_from_slice(&1u16.to_le_bytes());
-        bytes.extend_from_slice(&1u16.to_le_bytes());
-        bytes.extend_from_slice(&audio.sample_rate.to_le_bytes());
-        bytes.extend_from_slice(&byte_rate.to_le_bytes());
-        bytes.extend_from_slice(&2u16.to_le_bytes());
-        bytes.extend_from_slice(&16u16.to_le_bytes());
-        bytes.extend_from_slice(b"data");
-        bytes.extend_from_slice(&data_bytes.to_le_bytes());
-
-        for sample in &audio.samples {
-            let scaled = (sample.clamp(-1.0, 1.0) * i16::MAX as f32).round() as i16;
-            bytes.extend_from_slice(&scaled.to_le_bytes());
-        }
-
-        fs::write(path, bytes).map_err(|error| error.to_string())
-    }
-
-    fn read_wav_audio(path: &Path) -> Result<DecodedAudio, String> {
-        let bytes = fs::read(path).map_err(|error| error.to_string())?;
-        if bytes.len() < 12 || &bytes[0..4] != b"RIFF" || &bytes[8..12] != b"WAVE" {
-            return Err("Not a PCM WAV file.".to_string());
-        }
-
-        let mut audio_format = 0u16;
-        let mut channels = 0u16;
-        let mut sample_rate = 0u32;
-        let mut block_align = 0u16;
-        let mut bits_per_sample = 0u16;
-        let mut data_range: Option<(usize, usize)> = None;
-        let mut offset = 12usize;
-        while offset + 8 <= bytes.len() {
-            let chunk_id = &bytes[offset..offset + 4];
-            let chunk_size = read_u32_le(&bytes, offset + 4)
-                .ok_or_else(|| "Invalid WAV chunk header.".to_string())?
-                as usize;
-            let chunk_start = offset + 8;
-            let chunk_end = chunk_start.saturating_add(chunk_size);
-            if chunk_end > bytes.len() {
-                return Err("Invalid WAV chunk size.".to_string());
-            }
-            if chunk_id == b"fmt " {
-                if chunk_size < 16 {
-                    return Err("Invalid WAV fmt chunk.".to_string());
-                }
-                audio_format = read_u16_le(&bytes, chunk_start)
-                    .ok_or_else(|| "Invalid WAV audio format.".to_string())?;
-                channels = read_u16_le(&bytes, chunk_start + 2)
-                    .ok_or_else(|| "Invalid WAV channel count.".to_string())?;
-                sample_rate = read_u32_le(&bytes, chunk_start + 4)
-                    .ok_or_else(|| "Invalid WAV sample rate.".to_string())?;
-                block_align = read_u16_le(&bytes, chunk_start + 12)
-                    .ok_or_else(|| "Invalid WAV block alignment.".to_string())?;
-                bits_per_sample = read_u16_le(&bytes, chunk_start + 14)
-                    .ok_or_else(|| "Invalid WAV bit depth.".to_string())?;
-            } else if chunk_id == b"data" {
-                data_range = Some((chunk_start, chunk_end));
-            }
-            offset = chunk_end + (chunk_size % 2);
-        }
-
-        if audio_format != 1 && audio_format != 3 {
-            return Err("Lyrics WAV decode supports PCM and 32-bit float WAV files.".to_string());
-        }
-        if channels == 0 || sample_rate == 0 || block_align == 0 || bits_per_sample == 0 {
-            return Err("Invalid WAV stream metadata.".to_string());
-        }
-        let (data_start, data_end) =
-            data_range.ok_or_else(|| "WAV data chunk is missing.".to_string())?;
-        let frame_count = (data_end - data_start) / block_align as usize;
-        let bytes_per_sample = (bits_per_sample / 8) as usize;
-        if bytes_per_sample == 0 {
-            return Err("Invalid WAV bit depth.".to_string());
-        }
-
-        let mut samples = Vec::with_capacity(frame_count);
-        for frame_index in 0..frame_count {
-            let frame_start = data_start + frame_index * block_align as usize;
-            let mut sum = 0.0;
-            for channel_index in 0..channels as usize {
-                let sample_offset = frame_start + channel_index * bytes_per_sample;
-                sum += decode_wav_sample(&bytes, sample_offset, audio_format, bits_per_sample)?;
-            }
-            samples.push((sum / channels as f64).clamp(-1.0, 1.0) as f32);
-        }
-
-        Ok(DecodedAudio {
-            samples,
-            sample_rate,
-            channels: channels as u32,
-        })
-    }
-
-    struct MediaExtractorHandle {
-        ptr: *mut ndk_sys::AMediaExtractor,
-        data_source_file: Option<fs::File>,
-    }
-
-    impl MediaExtractorHandle {
-        fn new() -> Result<Self, String> {
-            let ptr = unsafe { ndk_sys::AMediaExtractor_new() };
-            if ptr.is_null() {
-                return Err("Android MediaExtractor could not be created.".to_string());
-            }
-            Ok(Self {
-                ptr,
-                data_source_file: None,
-            })
-        }
-
-        fn set_data_source(&mut self, path: &Path) -> Result<(), String> {
-            let file = fs::File::open(path).map_err(|error| {
-                format!("Android MediaExtractor could not open the imported audio: {error}")
-            })?;
-            let length: ndk_sys::off64_t = file
-                .metadata()
-                .map_err(|error| {
-                    format!("Android MediaExtractor could not inspect the imported audio: {error}")
-                })?
-                .len()
-                .try_into()
-                .map_err(|_| {
-                    "Imported audio is too large for Android MediaExtractor.".to_string()
-                })?;
-            let fd_status = unsafe {
-                ndk_sys::AMediaExtractor_setDataSourceFd(self.ptr, file.as_raw_fd(), 0, length)
-            };
-            if fd_status == ndk_sys::media_status_t::AMEDIA_OK {
-                self.data_source_file = Some(file);
-                return Ok(());
-            }
-
-            let path = CString::new(path.to_string_lossy().as_bytes())
-                .map_err(|_| "Audio path contains an invalid null byte.".to_string())?;
-            let status = unsafe { ndk_sys::AMediaExtractor_setDataSource(self.ptr, path.as_ptr()) };
-            if status != ndk_sys::media_status_t::AMEDIA_OK {
-                return Err("Android MediaExtractor could not open the imported audio.".to_string());
-            }
-            Ok(())
-        }
-    }
-
-    impl Drop for MediaExtractorHandle {
-        fn drop(&mut self) {
-            unsafe {
-                let _ = ndk_sys::AMediaExtractor_delete(self.ptr);
-            }
-        }
-    }
-
-    struct MediaFormatHandle(*mut ndk_sys::AMediaFormat);
-
-    impl MediaFormatHandle {
-        fn from_track(extractor: &MediaExtractorHandle, index: usize) -> Result<Self, String> {
-            let ptr = unsafe { ndk_sys::AMediaExtractor_getTrackFormat(extractor.ptr, index) };
-            if ptr.is_null() {
-                return Err("Android MediaExtractor returned an invalid track format.".to_string());
-            }
-            Ok(Self(ptr))
-        }
-    }
-
-    impl Drop for MediaFormatHandle {
-        fn drop(&mut self) {
-            unsafe {
-                let _ = ndk_sys::AMediaFormat_delete(self.0);
-            }
-        }
-    }
-
-    struct MediaCodecHandle {
-        ptr: *mut ndk_sys::AMediaCodec,
-        started: bool,
-    }
-
-    impl MediaCodecHandle {
-        fn new(mime: &str) -> Result<Self, String> {
-            let mime = CString::new(mime)
-                .map_err(|_| "Audio MIME type contains a null byte.".to_string())?;
-            let ptr = unsafe { ndk_sys::AMediaCodec_createDecoderByType(mime.as_ptr()) };
-            if ptr.is_null() {
-                return Err("Android MediaCodec has no decoder for this audio format.".to_string());
-            }
-            Ok(Self {
-                ptr,
-                started: false,
-            })
-        }
-
-        fn configure(&self, format: &MediaFormatHandle) -> Result<(), String> {
-            let status = unsafe {
-                ndk_sys::AMediaCodec_configure(
-                    self.ptr,
-                    format.0,
-                    ptr::null_mut(),
-                    ptr::null_mut(),
-                    0,
-                )
-            };
-            if status != ndk_sys::media_status_t::AMEDIA_OK {
-                return Err("Android MediaCodec could not configure the audio decoder.".to_string());
-            }
-            Ok(())
-        }
-
-        fn start(&mut self) -> Result<(), String> {
-            let status = unsafe { ndk_sys::AMediaCodec_start(self.ptr) };
-            if status != ndk_sys::media_status_t::AMEDIA_OK {
-                return Err("Android MediaCodec could not start the audio decoder.".to_string());
-            }
-            self.started = true;
-            Ok(())
-        }
-    }
-
-    impl Drop for MediaCodecHandle {
-        fn drop(&mut self) {
-            unsafe {
-                if self.started {
-                    let _ = ndk_sys::AMediaCodec_stop(self.ptr);
-                }
-                let _ = ndk_sys::AMediaCodec_delete(self.ptr);
-            }
-        }
-    }
-
-    fn media_format_string(
-        format: &MediaFormatHandle,
-        key: *const c_char,
-    ) -> Result<Option<String>, String> {
-        let mut value: *const c_char = ptr::null();
-        let found = unsafe { ndk_sys::AMediaFormat_getString(format.0, key, &mut value) };
-        if !found || value.is_null() {
-            return Ok(None);
-        }
-        let value = unsafe { CStr::from_ptr(value) }
-            .to_str()
-            .map_err(|_| "Android media format contained invalid UTF-8.".to_string())?
-            .to_string();
-        Ok(Some(value))
-    }
-
-    fn media_format_i32(format: *mut ndk_sys::AMediaFormat, key: *const c_char) -> Option<i32> {
-        let mut value = 0i32;
-        let found = unsafe { ndk_sys::AMediaFormat_getInt32(format, key, &mut value) };
-        found.then_some(value)
-    }
-
-    fn media_format_handle_i32(format: &MediaFormatHandle, key: *const c_char) -> Option<i32> {
-        media_format_i32(format.0, key)
-    }
-
-    fn read_android_media_audio(path: &Path) -> Result<DecodedAudio, String> {
-        const TIMEOUT_US: i64 = 10_000;
-        const PCM_ENCODING_16BIT: i32 = 2;
-        const PCM_ENCODING_FLOAT: i32 = 4;
-
-        let mut extractor = MediaExtractorHandle::new()?;
-        extractor.set_data_source(path)?;
-
-        let track_count = unsafe { ndk_sys::AMediaExtractor_getTrackCount(extractor.ptr) };
-        let mut selected_track = None;
-        let mut selected_format = None;
-        let mut selected_mime = None;
-        for track_index in 0..track_count {
-            let format = MediaFormatHandle::from_track(&extractor, track_index)?;
-            let mime = media_format_string(&format, unsafe { ndk_sys::AMEDIAFORMAT_KEY_MIME })?;
-            if mime
-                .as_deref()
-                .is_some_and(|value| value.starts_with("audio/"))
-            {
-                selected_track = Some(track_index);
-                selected_mime = mime;
-                selected_format = Some(format);
-                break;
-            }
-        }
-
-        let track_index = selected_track.ok_or_else(|| {
-            "Android MediaExtractor did not find an audio track in the imported file.".to_string()
-        })?;
-        let format = selected_format
-            .ok_or_else(|| "Android MediaExtractor lost the selected audio format.".to_string())?;
-        let mime = selected_mime.ok_or_else(|| {
-            "Android MediaExtractor did not report an audio MIME type.".to_string()
-        })?;
-
-        let sample_rate =
-            media_format_handle_i32(&format, unsafe { ndk_sys::AMEDIAFORMAT_KEY_SAMPLE_RATE })
-                .ok_or_else(|| {
-                    "Android MediaExtractor did not report an audio sample rate.".to_string()
-                })?;
-        let channels =
-            media_format_handle_i32(&format, unsafe { ndk_sys::AMEDIAFORMAT_KEY_CHANNEL_COUNT })
-                .ok_or_else(|| {
-                    "Android MediaExtractor did not report an audio channel count.".to_string()
-                })?;
-        if sample_rate <= 0 || channels <= 0 {
-            return Err("Android MediaExtractor reported invalid audio metadata.".to_string());
-        }
-
-        let status = unsafe { ndk_sys::AMediaExtractor_selectTrack(extractor.ptr, track_index) };
-        if status != ndk_sys::media_status_t::AMEDIA_OK {
-            return Err("Android MediaExtractor could not select the audio track.".to_string());
-        }
-
-        let mut codec = MediaCodecHandle::new(&mime)?;
-        codec.configure(&format)?;
-        codec.start()?;
-
-        let mut output_sample_rate = sample_rate;
-        let mut output_channels = channels;
-        let mut output_encoding = PCM_ENCODING_16BIT;
-        let mut decoded = Vec::new();
-        let mut input_eos = false;
-        let mut output_eos = false;
-        let mut idle_iterations = 0usize;
-
-        while !output_eos {
-            let mut made_progress = false;
-            if !input_eos {
-                let input_index =
-                    unsafe { ndk_sys::AMediaCodec_dequeueInputBuffer(codec.ptr, TIMEOUT_US) };
-                if input_index >= 0 {
-                    let mut input_capacity = 0usize;
-                    let input_buffer = unsafe {
-                        ndk_sys::AMediaCodec_getInputBuffer(
-                            codec.ptr,
-                            input_index as usize,
-                            &mut input_capacity,
-                        )
-                    };
-                    if input_buffer.is_null() {
-                        return Err(
-                            "Android MediaCodec returned an invalid input buffer.".to_string()
-                        );
-                    }
-                    let sample_size = unsafe {
-                        ndk_sys::AMediaExtractor_readSampleData(
-                            extractor.ptr,
-                            input_buffer,
-                            input_capacity,
-                        )
-                    };
-                    if sample_size < 0 {
-                        let status = unsafe {
-                            ndk_sys::AMediaCodec_queueInputBuffer(
-                                codec.ptr,
-                                input_index as usize,
-                                0,
-                                0,
-                                0,
-                                ndk_sys::AMEDIACODEC_BUFFER_FLAG_END_OF_STREAM,
-                            )
-                        };
-                        if status != ndk_sys::media_status_t::AMEDIA_OK {
-                            return Err("Android MediaCodec could not queue audio end-of-stream."
-                                .to_string());
-                        }
-                        input_eos = true;
-                    } else {
-                        let sample_time =
-                            unsafe { ndk_sys::AMediaExtractor_getSampleTime(extractor.ptr) };
-                        let sample_flags =
-                            unsafe { ndk_sys::AMediaExtractor_getSampleFlags(extractor.ptr) };
-                        let status = unsafe {
-                            ndk_sys::AMediaCodec_queueInputBuffer(
-                                codec.ptr,
-                                input_index as usize,
-                                0,
-                                sample_size as usize,
-                                sample_time.max(0) as u64,
-                                sample_flags,
-                            )
-                        };
-                        if status != ndk_sys::media_status_t::AMEDIA_OK {
-                            return Err(
-                                "Android MediaCodec could not queue compressed audio.".to_string()
-                            );
-                        }
-                        unsafe {
-                            ndk_sys::AMediaExtractor_advance(extractor.ptr);
-                        }
-                    }
-                    made_progress = true;
-                }
-            }
-
-            let mut info: ndk_sys::AMediaCodecBufferInfo = unsafe { mem::zeroed() };
-            let output_index = unsafe {
-                ndk_sys::AMediaCodec_dequeueOutputBuffer(codec.ptr, &mut info, TIMEOUT_US)
-            };
-            if output_index >= 0 {
-                let mut output_capacity = 0usize;
-                let output_buffer = unsafe {
-                    ndk_sys::AMediaCodec_getOutputBuffer(
-                        codec.ptr,
-                        output_index as usize,
-                        &mut output_capacity,
-                    )
-                };
-                if output_buffer.is_null() {
-                    return Err("Android MediaCodec returned an invalid output buffer.".to_string());
-                }
-                let offset = info.offset.max(0) as usize;
-                let size = info.size.max(0) as usize;
-                if size > 0 {
-                    let end = offset.saturating_add(size);
-                    if end > output_capacity {
-                        return Err(
-                            "Android MediaCodec returned an invalid output range.".to_string()
-                        );
-                    }
-                    let output = unsafe { slice::from_raw_parts(output_buffer.add(offset), size) };
-                    decoded.extend_from_slice(output);
-                }
-                if info.flags & ndk_sys::AMEDIACODEC_BUFFER_FLAG_END_OF_STREAM != 0 {
-                    output_eos = true;
-                }
-                unsafe {
-                    ndk_sys::AMediaCodec_releaseOutputBuffer(
-                        codec.ptr,
-                        output_index as usize,
-                        false,
-                    );
-                }
-                made_progress = true;
-            } else if output_index == ndk_sys::AMEDIACODEC_INFO_OUTPUT_FORMAT_CHANGED as isize {
-                let output_format = unsafe { ndk_sys::AMediaCodec_getOutputFormat(codec.ptr) };
-                if !output_format.is_null() {
-                    output_sample_rate = media_format_i32(output_format, unsafe {
-                        ndk_sys::AMEDIAFORMAT_KEY_SAMPLE_RATE
-                    })
-                    .unwrap_or(output_sample_rate);
-                    output_channels = media_format_i32(output_format, unsafe {
-                        ndk_sys::AMEDIAFORMAT_KEY_CHANNEL_COUNT
-                    })
-                    .unwrap_or(output_channels);
-                    output_encoding = media_format_i32(output_format, unsafe {
-                        ndk_sys::AMEDIAFORMAT_KEY_PCM_ENCODING
-                    })
-                    .unwrap_or(PCM_ENCODING_16BIT);
-                    unsafe {
-                        let _ = ndk_sys::AMediaFormat_delete(output_format);
-                    }
-                }
-                made_progress = true;
-            } else if output_index != ndk_sys::AMEDIACODEC_INFO_TRY_AGAIN_LATER as isize {
-                return Err("Android MediaCodec failed while decoding audio.".to_string());
-            }
-
-            if made_progress {
-                idle_iterations = 0;
-            } else {
-                idle_iterations += 1;
-                if idle_iterations > 2_000 {
-                    return Err("Android MediaCodec timed out while decoding audio.".to_string());
-                }
-            }
-        }
-
-        match output_encoding {
-            PCM_ENCODING_16BIT => {
-                pcm_i16_bytes_to_mono(&decoded, output_sample_rate, output_channels)
-            }
-            PCM_ENCODING_FLOAT => {
-                pcm_f32_bytes_to_mono(&decoded, output_sample_rate, output_channels)
-            }
-            _ => Err("Android MediaCodec returned an unsupported PCM output encoding.".to_string()),
-        }
-    }
-
-    fn pcm_i16_bytes_to_mono(
-        bytes: &[u8],
-        sample_rate: i32,
-        channels: i32,
-    ) -> Result<DecodedAudio, String> {
-        if sample_rate <= 0 || channels <= 0 {
-            return Err("Android MediaCodec returned invalid PCM metadata.".to_string());
-        }
-        let channel_count = channels as usize;
-        let frame_bytes = channel_count
-            .checked_mul(2)
-            .ok_or_else(|| "Android MediaCodec returned invalid channel metadata.".to_string())?;
-        if frame_bytes == 0 {
-            return Err("Android MediaCodec returned invalid channel metadata.".to_string());
-        }
-        let frame_count = bytes.len() / frame_bytes;
-        let mut samples = Vec::with_capacity(frame_count);
-        for frame_index in 0..frame_count {
-            let frame_start = frame_index * frame_bytes;
-            let mut sum = 0.0f32;
-            for channel_index in 0..channel_count {
-                let sample_start = frame_start + channel_index * 2;
-                let raw = i16::from_le_bytes(
-                    bytes
-                        .get(sample_start..sample_start + 2)
-                        .ok_or_else(|| {
-                            "Android MediaCodec returned truncated PCM data.".to_string()
-                        })?
-                        .try_into()
-                        .map_err(|_| "Android MediaCodec returned invalid PCM data.".to_string())?,
-                );
-                sum += raw as f32 / i16::MAX as f32;
-            }
-            samples.push((sum / channel_count as f32).clamp(-1.0, 1.0));
-        }
-        Ok(DecodedAudio {
-            samples,
-            sample_rate: sample_rate as u32,
-            channels: channels as u32,
-        })
-    }
-
-    fn pcm_f32_bytes_to_mono(
-        bytes: &[u8],
-        sample_rate: i32,
-        channels: i32,
-    ) -> Result<DecodedAudio, String> {
-        if sample_rate <= 0 || channels <= 0 {
-            return Err("Android MediaCodec returned invalid PCM metadata.".to_string());
-        }
-        let channel_count = channels as usize;
-        let frame_bytes = channel_count
-            .checked_mul(4)
-            .ok_or_else(|| "Android MediaCodec returned invalid channel metadata.".to_string())?;
-        if frame_bytes == 0 {
-            return Err("Android MediaCodec returned invalid channel metadata.".to_string());
-        }
-        let frame_count = bytes.len() / frame_bytes;
-        let mut samples = Vec::with_capacity(frame_count);
-        for frame_index in 0..frame_count {
-            let frame_start = frame_index * frame_bytes;
-            let mut sum = 0.0f32;
-            for channel_index in 0..channel_count {
-                let sample_start = frame_start + channel_index * 4;
-                let raw = f32::from_le_bytes(
-                    bytes
-                        .get(sample_start..sample_start + 4)
-                        .ok_or_else(|| {
-                            "Android MediaCodec returned truncated PCM data.".to_string()
-                        })?
-                        .try_into()
-                        .map_err(|_| "Android MediaCodec returned invalid PCM data.".to_string())?,
-                );
-                sum += raw;
-            }
-            samples.push((sum / channel_count as f32).clamp(-1.0, 1.0));
-        }
-        Ok(DecodedAudio {
-            samples,
-            sample_rate: sample_rate as u32,
-            channels: channels as u32,
-        })
-    }
-
-    fn read_lyrics_audio(path: &Path) -> Result<DecodedAudio, String> {
-        let decoded = read_mobile_audio(path)?;
-        Ok(DecodedAudio {
-            samples: resample_mono(&decoded.samples, decoded.sample_rate, WHISPER_SAMPLE_RATE),
-            sample_rate: WHISPER_SAMPLE_RATE,
-            channels: 1,
-        })
-    }
-
-    fn resample_mono(
-        samples: &[f32],
-        source_sample_rate: u32,
-        target_sample_rate: u32,
-    ) -> Vec<f32> {
-        if samples.is_empty() || source_sample_rate == 0 || source_sample_rate == target_sample_rate
-        {
-            return samples.to_vec();
-        }
-
-        let output_len = ((samples.len() as f64) * target_sample_rate as f64
-            / source_sample_rate as f64)
-            .ceil()
-            .max(1.0) as usize;
-        let ratio = source_sample_rate as f64 / target_sample_rate as f64;
-        let mut output = Vec::with_capacity(output_len);
-        for index in 0..output_len {
-            let source_position = index as f64 * ratio;
-            let left_index = source_position.floor() as usize;
-            let right_index = (left_index + 1).min(samples.len().saturating_sub(1));
-            let fraction = (source_position - left_index as f64) as f32;
-            let left = samples[left_index];
-            let right = samples[right_index];
-            output.push(left + (right - left) * fraction);
-        }
-        output
-    }
-
-    fn read_u16_le(bytes: &[u8], offset: usize) -> Option<u16> {
-        Some(u16::from_le_bytes(
-            bytes.get(offset..offset + 2)?.try_into().ok()?,
-        ))
-    }
-
-    fn read_u32_le(bytes: &[u8], offset: usize) -> Option<u32> {
-        Some(u32::from_le_bytes(
-            bytes.get(offset..offset + 4)?.try_into().ok()?,
-        ))
-    }
-
-    fn decode_wav_sample(
-        bytes: &[u8],
-        offset: usize,
-        audio_format: u16,
-        bits_per_sample: u16,
-    ) -> Result<f64, String> {
-        match (audio_format, bits_per_sample) {
-            (1, 8) => Ok((bytes
-                .get(offset)
-                .copied()
-                .ok_or_else(|| "Invalid WAV sample data.".to_string())?
-                as f64
-                - 128.0)
-                / 128.0),
-            (1, 16) => {
-                let raw = i16::from_le_bytes(
-                    bytes
-                        .get(offset..offset + 2)
-                        .ok_or_else(|| "Invalid WAV sample data.".to_string())?
-                        .try_into()
-                        .map_err(|_| "Invalid WAV sample data.".to_string())?,
-                );
-                Ok(raw as f64 / i16::MAX as f64)
-            }
-            (1, 24) => {
-                let sample = bytes
-                    .get(offset..offset + 3)
-                    .ok_or_else(|| "Invalid WAV sample data.".to_string())?;
-                let raw =
-                    (sample[0] as i32) | ((sample[1] as i32) << 8) | ((sample[2] as i32) << 16);
-                let signed = if raw & 0x800000 != 0 {
-                    raw | !0x00ff_ffff
-                } else {
-                    raw
-                };
-                Ok(signed as f64 / 8_388_608.0)
-            }
-            (1, 32) => {
-                let raw = i32::from_le_bytes(
-                    bytes
-                        .get(offset..offset + 4)
-                        .ok_or_else(|| "Invalid WAV sample data.".to_string())?
-                        .try_into()
-                        .map_err(|_| "Invalid WAV sample data.".to_string())?,
-                );
-                Ok(raw as f64 / i32::MAX as f64)
-            }
-            (3, 32) => {
-                let raw = f32::from_le_bytes(
-                    bytes
-                        .get(offset..offset + 4)
-                        .ok_or_else(|| "Invalid WAV sample data.".to_string())?
-                        .try_into()
-                        .map_err(|_| "Invalid WAV sample data.".to_string())?,
-                );
-                Ok(raw as f64)
-            }
-            _ => Err(
-                "Mobile CPU analysis currently supports 8/16/24/32-bit PCM WAV files.".to_string(),
-            ),
-        }
     }
 
     fn pitch_class_energy(samples: &[f64], sample_rate: f64) -> [f64; 12] {
@@ -2179,7 +1468,7 @@ mod android {
         source_path: &Path,
         model: &WhisperModel,
     ) -> Result<MobileLyricsTranscription, String> {
-        let audio = read_lyrics_audio(source_path)?;
+        let audio = read_resampled_mono_audio(source_path, WHISPER_SAMPLE_RATE)?;
         if audio.samples.is_empty() {
             return Err(
                 "Imported audio did not contain samples for lyrics transcription.".to_string(),
